Feat: get zkevmsha256 working WIP
